pyproject.toml
[project]
name = "chops-tui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "AYIXIAYI_X", email = "ayixiyouxiang@163.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "textual>=8.1.1",
    "watchdog>=6.0.0",
]

[project.scripts]
chops-tui = "chops_tui:main"

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"
